// src/app/risk/model/riskDashboard.ts
export class RiskDashboard {
  Source: string;
  RiskRequestId: number;
  HazardIdentified: string;
  RiskIdentified: string;
  PotentialRiskPear: string;
  PotentialRiskLikelihood: string;
  PotentialRiskSeverity: number;
  CreatedDate: Date;
  NoOfActionsAssigned: number;
  NoOfActionCompleted: number;
  NoOfActionPending: number;
  ResidualRiskLikelihood: string;
  ResidualRiskSeverity: number;
  ResidualRisk: string;
}